pyproject.toml
[tool.poetry]
name = "encodeck"
version = "0.1.0"
description = ""
authors = ["dhyutitewani <dhyuti07@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
pycryptodome = "^3.21.0"
decouple = "^0.0.7"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
